Adds skip root deployment and version warning
Allows skipping deployment of the root package and disables the version check warning.

This change introduces a `skipRoot` configuration option in `config.toml`.
If set to true, the root package will be excluded from the deployment process.

Adds a `--skip-version-warning` flag to suppress version update notifications.
Also, version warning check is now skipped in `PersistentPreRunE` if flag is set.